@@ -1,5 +1,5 @@
-using ManyConsole;
 using MaiLib;
+using ManyConsole;
 
 namespace MaichartConverter
 {
@@ -90,16 +90,17 @@ namespace MaichartConverter
         {
             try
             {
-                Ma2Tokenizer tokenizer = new Ma2Tokenizer();
-                Ma2Parser parser = new Ma2Parser();
+                Ma2Tokenizer tokenizer = new();
+                Ma2Parser parser = new();
                 //Chart good = new Ma2(@"/Users/neskol/MaiAnalysis/A000/music/music" + musicID + "/" + musicID + "_0" + difficulty + ".ma2");
                 string tokenLocation = FileLocation ?? throw new FileNotFoundException("Token location not specified");
-                string compensatedId = Program.CompensateZero(ID ?? throw new NullReferenceException("ID shall not be null"));
-                
+                string compensatedId =
+                    Program.CompensateZero(ID ?? throw new NullReferenceException("ID shall not be null"));
+
                 string ma2Path = Path.Combine(
-                    tokenLocation, 
-                    "music", 
-                    $"music{compensatedId}", 
+                    tokenLocation,
+                    "music",
+                    $"music{compensatedId}",
                     $"{compensatedId}_0{Difficulty}.ma2"
                 );
 
@@ -118,10 +119,7 @@ namespace MaichartConverter
                     candidate.RotateNotes(rotateMethod);
                 }
 
-                if (ShiftTick != null && ShiftTick != 0)
-                {
-                    candidate.ShiftByOffset((int)ShiftTick);
-                }
+                if (ShiftTick != null && ShiftTick != 0) candidate.ShiftByOffset((int)ShiftTick);
 
                 string result = "";
                 switch (TargetFormat)
@@ -131,28 +129,27 @@ namespace MaichartConverter
                     case "simai":
                     case "Simai":
                     case "SimaiFes":
-                        Simai resultChart = new Simai(candidate);
+                        Simai resultChart = new(candidate);
                         // resultChart.Update();
                         result = resultChart.Compose();
                         if (Destination != null && !Destination.Equals(""))
                         {
                             string targetMaidataLocation = $"{Destination}/maidata.txt";
                             if (!Directory.Exists(Destination)) Directory.CreateDirectory(Destination);
-                            StreamWriter sw = new StreamWriter(targetMaidataLocation, false);
+                            StreamWriter sw = new(targetMaidataLocation, false);
                             {
                                 sw.WriteLine(result);
                             }
                             sw.Close();
                             if (File.Exists(targetMaidataLocation))
-                            {
                                 Console.WriteLine("Successfully compiled at: {0}", targetMaidataLocation);
-                            }
                             else
-                            {
                                 throw new FileNotFoundException("THE FILE IS NOT SUCCESSFULLY COMPILED.");
-                            }
                         }
-                        else Console.WriteLine(result);
+                        else
+                        {
+                            Console.WriteLine(result);
+                        }
 
                         break;
                     case "ma2":
@@ -172,21 +169,20 @@ namespace MaichartConverter
                         {
                             string targetMaidataLocation = $"{Destination}/result.ma2";
                             if (!Directory.Exists(Destination)) Directory.CreateDirectory(Destination);
-                            StreamWriter sw = new StreamWriter(targetMaidataLocation, false);
+                            StreamWriter sw = new(targetMaidataLocation, false);
                             {
                                 sw.WriteLine(result);
                             }
                             sw.Close();
                             if (File.Exists(targetMaidataLocation))
-                            {
                                 Console.WriteLine("Successfully compiled at: {0}", targetMaidataLocation);
-                            }
                             else
-                            {
                                 throw new FileNotFoundException("THE FILE IS NOT SUCCESSFULLY COMPILED.");
-                            }
                         }
-                        else Console.WriteLine(result);
+                        else
+                        {
+                            Console.WriteLine(result);
+                        }
 
                         break;
                 }
